Rename register_observable_param() to register_observable().

+++ tests/unit/lib/parameter_observers/ObservedValue_unittest.cc
@@ -13,5 +13,5 @@ TEST(ObservedValue, set_correct)
 	auto obs = ObservedValue::make_observation(1, "test", 42);
 	EXPECT_EQ(obs->get<int64_t>("step"), 1);
 	EXPECT_EQ(obs->get<std::string>("name"), "test");
-	EXPECT_EQ(obs->get<int64_t>("value"), 42);
+	EXPECT_EQ(obs->get<int32_t>("value"), 42);
 };
